+++ pybmd/__init__.py
@@ -4,8 +4,8 @@ from .ui_dispather import UI_Dispather
 __name__ = 'pybmd'
 
 __major_version__ = "2025"
-__minor_version__ = "1"
-__change_version__ = "1"
+__minor_version__ = "2"
+__change_version__ = "0"
 __version__ = ".".join(
     (__major_version__, __minor_version__, __change_version__)
 )